requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.2,<25.0
psycopg2-binary>=2.9,<3.0

# Environment variables
python-dotenv>=1.0,<2.0

# Auth (JWT)
djangorestframework-simplejwt>=5.3,<6.0

# WebSocket support (Channels)
channels>=4.3,<5.0
channels-redis>=4.3,<5.0

# CORS support
django-cors-headers>=4.3,<5.0

# API docs (optional but recommended)
drf-spectacular>=0.27,<0.28

# Developer tools
black>=24.4,<25.0
isort>=5.13,<6.0
flake8>=7.1,<8.0
pre-commit>=3.8,<4.0
